+++ src/components/common/Button/index.tsx
@@ -1,0 +1,40 @@
+import { ReactComponent as Add } from '@assets/add.svg';
+
+import { CommonButton } from './Button.style';
+import { ButtonProps } from './Button.types';
+
+/**
+ * Button component
+ * @param size - 사이즈를 나타냅니다. 'small' | 'medium'  기본은 'medium' 입니다.
+ * @param variant - style을 나타냅니다.
+ * @param hasIcon - icon 존재 여부를 나타냅니다.
+ * @param isDisabled - 버튼 비활성화 여부를 나타냅니다.
+ * @param iconPosition - icon 위치를 나타냅니다. 'left' | 'right' 기본은 'left' 입니다.
+ * @param onClick - 클릭 이벤트를 나타냅니다.
+ * @param text - 버튼 텍스트를 나타냅니다.
+ */
+
+export const Button = ({
+  size = 'medium',
+  variant,
+  hasIcon = true,
+  isDisabled = false,
+  iconPosition = 'left',
+  onClick,
+  text,
+}: ButtonProps) => (
+  <CommonButton
+    size={size}
+    variant={variant}
+    disabled={isDisabled}
+    onClick={onClick}
+  >
+    {hasIcon && iconPosition === 'left' && (
+      <Add fill="current" color="current" />
+    )}
+    {text}
+    {hasIcon && iconPosition === 'right' && (
+      <Add fill="current" color="current" />
+    )}
+  </CommonButton>
+);
